requirements.txt
flask
graphene==2.1.9
requests
pyjwt
cryptography
pyyaml